#1 - Allow Cheking and retrieving of permissions for android and iOS
Note:: For iOS there is no Calendar Read vs Write permissions, instead they have Calendar and Reminder permissions, which are quite different.
https://docs.microsoft.com/en-us/xamarin/ios/platform/eventkit

I've also noticed that although their is two seperate permissions for Calendar Read/Write for Android, when looking at the settings for the app after providing it with Read/Write Access it seems to set both as active, as there is no seperation in switches.