Using new UWP recurrence rule model
(with a readonly version existing to either allow us to view rules we do not process the power to create, or to revert back to at a later stage)

the iOS and Android is commented out at this stage as they will most likely require reworks once we have a clear vision going forward.

the selection of days of the week works for selection but not initilization will be looking to convert to a list of checkboxes tomorrow if I am unable to solve it within the first hour.